requirements.txt
GitPython==3.1.2
launchpadlib==1.10.13
progress==1.5
